backend/requirements.txt
# Django CMS 后端项目依赖
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
